+++ server/src/test/java/org/opensearch/cluster/coordination/CoordinationStateTests.java
@@ -938,6 +938,7 @@ public class CoordinationStateTests extends OpenSearchTestCase {
             Version.CURRENT,
             randomAlphaOfLength(10),
             false,
+            1,
             randomAlphaOfLength(10),
             Collections.emptyList(),
             randomAlphaOfLength(10),